engine: Enable image cleanup tests for Windows
