+++ agent/engine/docker_image_manager_unix_integ_test.go
@@ -1,4 +1,4 @@
-// +build integration
+// +build !windows,integration
 // Copyright 2014-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
@@ -18,8 +18,6 @@
 package engine
 
 import (
-	"container/list"
-	"errors"
 	"fmt"
 	"testing"
 	"time"
@@ -31,249 +29,20 @@ import (
 )
 
 const (
-	imageRemovalTimeout       = 30 * time.Second
-	taskCleanupTimeoutSeconds = 30
+	test1Image1Name = "127.0.0.1:51670/amazon/image-cleanup-test-image1:latest"
+	test1Image2Name = "127.0.0.1:51670/amazon/image-cleanup-test-image2:latest"
+	test1Image3Name = "127.0.0.1:51670/amazon/image-cleanup-test-image3:latest"
 
-	testImage1Name = "127.0.0.1:51670/amazon/image-cleanup-test-image1:latest"
-	testImage2Name = "127.0.0.1:51670/amazon/image-cleanup-test-image2:latest"
-	testImage3Name = "127.0.0.1:51670/amazon/image-cleanup-test-image3:latest"
+	test2Image1Name = "127.0.0.1:51670/amazon/image-cleanup-test-image1:latest"
+	test2Image2Name = "127.0.0.1:51670/amazon/image-cleanup-test-image2:latest"
+	test2Image3Name = "127.0.0.1:51670/amazon/image-cleanup-test-image3:latest"
+
+	test3Image1Name = "127.0.0.1:51670/amazon/image-cleanup-test-image1:latest"
+	test3Image2Name = "127.0.0.1:51670/amazon/image-cleanup-test-image2:latest"
+	test3Image3Name = "127.0.0.1:51670/amazon/image-cleanup-test-image3:latest"
+
+	test4Image1Name = "127.0.0.1:51670/amazon/image-cleanup-test-image1:latest"
 )
-
-// Deletion of images in the order of LRU time: Happy path
-//  a. This includes starting up agent, pull images, start containers,
-//    account them in image manager,  stop containers, remove containers, account this in image manager,
-//  b. Simulate the pulled time (so that it passes the minimum age criteria
-//    for getting chosen for deletion )
-//  c. Start image cleanup , ensure that ONLY the top 2 eligible LRU images
-//    are removed from the instance,  and those deleted images’ image states are removed from image manager.
-//  d. Ensure images that do not pass the ‘minimumAgeForDeletion’ criteria are not removed.
-//  e. Image has not passed the ‘hasNoAssociatedContainers’ criteria.
-//  f. Ensure that that if not eligible, image is not deleted from the instance and image reference in ImageManager is not removed.
-func TestIntegImageCleanupHappyCase(t *testing.T) {
-	cfg := defaultTestConfigIntegTest()
-	cfg.TaskCleanupWaitDuration = 5 * time.Second
-
-	// Set low values so this test can complete in a sane amout of time
-	cfg.MinimumImageDeletionAge = 1 * time.Second
-	cfg.NumImagesToDeletePerCycle = 2
-	// start agent
-	taskEngine, done, _ := setup(cfg, t)
-
-	imageManager := taskEngine.(*DockerTaskEngine).imageManager.(*dockerImageManager)
-	imageManager.SetSaver(statemanager.NewNoopStateManager())
-
-	defer func() {
-		done()
-		// Force cleanup all test images and containers
-		cleanupImages(imageManager)
-	}()
-
-	taskEvents, containerEvents := taskEngine.TaskEvents()
-
-	defer discardEvents(containerEvents)()
-
-	// Create test Task
-	taskName := "imgClean"
-	testTask := createImageCleanupTestTask(taskName)
-
-	go taskEngine.AddTask(testTask)
-
-	// Verify that Task is running
-	err := verifyTaskIsRunning(taskEvents, testTask)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	imageState1 := imageManager.GetImageStateFromImageName(testImage1Name)
-	if imageState1 == nil {
-		t.Fatalf("Could not find image state for %s", testImage1Name)
-	} else {
-		t.Logf("Found image state for %s", testImage1Name)
-	}
-	imageState2 := imageManager.GetImageStateFromImageName(testImage2Name)
-	if imageState2 == nil {
-		t.Fatalf("Could not find image state for %s", testImage2Name)
-	} else {
-		t.Logf("Found image state for %s", testImage2Name)
-	}
-	imageState3 := imageManager.GetImageStateFromImageName(testImage3Name)
-	if imageState3 == nil {
-		t.Fatalf("Could not find image state for %s", testImage3Name)
-	} else {
-		t.Logf("Found image state for %s", testImage3Name)
-	}
-
-	imageState1ImageID := imageState1.Image.ImageID
-	imageState2ImageID := imageState2.Image.ImageID
-	imageState3ImageID := imageState3.Image.ImageID
-
-	// Set the ImageState.LastUsedAt to a value far in the past to ensure the test images are deleted.
-	// This will make these test images the LRU images.
-	imageState1.LastUsedAt = imageState1.LastUsedAt.Add(-99995 * time.Hour)
-	imageState2.LastUsedAt = imageState2.LastUsedAt.Add(-99994 * time.Hour)
-	imageState3.LastUsedAt = imageState3.LastUsedAt.Add(-99993 * time.Hour)
-
-	// Verify Task is stopped.
-	verifyTaskIsStopped(taskEvents, testTask)
-
-	// Allow Task cleanup to occur
-	time.Sleep(5 * time.Second)
-
-	// Verify Task is cleaned up
-	err = verifyTaskIsCleanedUp(taskName, taskEngine)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Call Image removal
-	imageManager.removeUnusedImages()
-
-	// Verify top 2 LRU images are deleted from image manager
-	err = verifyImagesAreRemoved(imageManager, imageState1ImageID, imageState2ImageID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify 3rd LRU image is not removed
-	err = verifyImagesAreNotRemoved(imageManager, imageState3ImageID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify top 2 LRU images are removed from docker
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState1ImageID)
-	if err != docker.ErrNoSuchImage {
-		t.Fatalf("Image was not removed successfully")
-	}
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState2ImageID)
-	if err != docker.ErrNoSuchImage {
-		t.Fatalf("Image was not removed successfully")
-	}
-
-	// Verify 3rd LRU image has not been removed from Docker
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState3ImageID)
-	if err != nil {
-		t.Fatalf("Image should not have been removed from Docker")
-	}
-}
-
-// Test that images not falling in the image deletion eligibility criteria are not removed:
-//  a. Ensure images that do not pass the ‘minimumAgeForDeletion’ criteria are not removed.
-//  b. Image has not passed the ‘hasNoAssociatedContainers’ criteria.
-//  c. Ensure that the image is not deleted from the instance and image reference in ImageManager is not removed.
-func TestIntegImageCleanupThreshold(t *testing.T) {
-	cfg := defaultTestConfigIntegTest()
-	cfg.TaskCleanupWaitDuration = 1 * time.Second
-
-	// Set low values so this test can complete in a sane amout of time
-	cfg.MinimumImageDeletionAge = 15 * time.Minute
-	// Set to delete three images, but in this test we expect only two images to be removed
-	cfg.NumImagesToDeletePerCycle = 3
-	// start agent
-	taskEngine, done, _ := setup(cfg, t)
-
-	imageManager := taskEngine.(*DockerTaskEngine).imageManager.(*dockerImageManager)
-	imageManager.SetSaver(statemanager.NewNoopStateManager())
-
-	defer func() {
-		done()
-		// Force cleanup all test images and containers
-		cleanupImages(imageManager)
-	}()
-
-	taskEvents, containerEvents := taskEngine.TaskEvents()
-	defer discardEvents(containerEvents)()
-
-	// Create test Task
-	taskName := "imgClean"
-	testTask := createImageCleanupTestTask(taskName)
-
-	// Start Task
-	go taskEngine.AddTask(testTask)
-
-	// Verify that Task is running
-	err := verifyTaskIsRunning(taskEvents, testTask)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	imageState1 := imageManager.GetImageStateFromImageName(testImage1Name)
-	if imageState1 == nil {
-		t.Fatalf("Could not find image state for %s", testImage1Name)
-	} else {
-		t.Logf("Found image state for %s", testImage1Name)
-	}
-	imageState2 := imageManager.GetImageStateFromImageName(testImage2Name)
-	if imageState2 == nil {
-		t.Fatalf("Could not find image state for %s", testImage2Name)
-	} else {
-		t.Logf("Found image state for %s", testImage2Name)
-	}
-	imageState3 := imageManager.GetImageStateFromImageName(testImage3Name)
-	if imageState3 == nil {
-		t.Fatalf("Could not find image state for %s", testImage3Name)
-	} else {
-		t.Logf("Found image state for %s", testImage3Name)
-	}
-
-	imageState1ImageID := imageState1.Image.ImageID
-	imageState2ImageID := imageState2.Image.ImageID
-	imageState3ImageID := imageState3.Image.ImageID
-
-	// Set the ImageState.LastUsedAt to a value far in the past to ensure the test images are deleted.
-	// This will make these the LRU images so they are deleted.
-	imageState1.LastUsedAt = imageState1.LastUsedAt.Add(-99995 * time.Hour)
-	imageState2.LastUsedAt = imageState2.LastUsedAt.Add(-99994 * time.Hour)
-	imageState3.LastUsedAt = imageState3.LastUsedAt.Add(-99993 * time.Hour)
-
-	// Set two containers to have pull time > threshold
-	imageState1.PulledAt = imageState1.PulledAt.Add(-20 * time.Minute)
-	imageState2.PulledAt = imageState2.PulledAt.Add(-10 * time.Minute)
-	imageState3.PulledAt = imageState3.PulledAt.Add(-25 * time.Minute)
-
-	// Verify Task is stopped
-	verifyTaskIsStopped(taskEvents, testTask)
-
-	// Allow Task cleanup to occur
-	time.Sleep(2 * time.Second)
-
-	// Verify Task is cleaned up
-	err = verifyTaskIsCleanedUp(taskName, taskEngine)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Call Image removal
-	imageManager.removeUnusedImages()
-
-	// Verify Image1 & Image3 are removed from ImageManager as they are beyond the minimumAge threshold
-	err = verifyImagesAreRemoved(imageManager, imageState1ImageID, imageState3ImageID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify Image2 is not removed, below threshold for minimumAge
-	err = verifyImagesAreNotRemoved(imageManager, imageState2ImageID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify Image1 & Image3 are removed from docker
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState1ImageID)
-	if err != docker.ErrNoSuchImage {
-		t.Fatalf("Image was not removed successfully")
-	}
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState3ImageID)
-	if err != docker.ErrNoSuchImage {
-		t.Fatalf("Image was not removed successfully")
-	}
-
-	// Verify Image2 has not been removed from Docker
-	_, err = taskEngine.(*DockerTaskEngine).client.InspectImage(imageState2ImageID)
-	if err != nil {
-		t.Fatalf("Image should not have been removed from Docker")
-	}
-}
 
 // TestImageWithSameNameAndDifferentID tests image can be correctly removed when tasks
 // are running with the same image name, but different image id.
@@ -300,17 +69,17 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	defer discardEvents(containerEvents)()
 
 	// Pull the images needed for the test
-	if _, err := dockerClient.InspectImage(testImage1Name); err == docker.ErrNoSuchImage {
-		metadata := dockerClient.PullImage(testImage1Name, nil)
-		assert.NoError(t, metadata.Error, "Failed to pull image %s", testImage1Name)
+	if _, err = dockerClient.InspectImage(test3Image1Name); err == docker.ErrNoSuchImage {
+		metadata := dockerClient.PullImage(test3Image1Name, nil)
+		assert.NoError(t, metadata.Error, "Failed to pull image %s", test3Image1Name)
 	}
-	if _, err := dockerClient.InspectImage(testImage2Name); err == docker.ErrNoSuchImage {
-		metadata := dockerClient.PullImage(testImage2Name, nil)
-		assert.NoError(t, metadata.Error, "Failed to pull image %s", testImage2Name)
+	if _, err = dockerClient.InspectImage(test3Image2Name); err == docker.ErrNoSuchImage {
+		metadata := dockerClient.PullImage(test3Image2Name, nil)
+		assert.NoError(t, metadata.Error, "Failed to pull image %s", test3Image2Name)
 	}
-	if _, err := dockerClient.InspectImage(testImage3Name); err == docker.ErrNoSuchImage {
-		metadata := dockerClient.PullImage(testImage3Name, nil)
-		assert.NoError(t, metadata.Error, "Failed to pull image %s", testImage3Name)
+	if _, err = dockerClient.InspectImage(test3Image3Name); err == docker.ErrNoSuchImage {
+		metadata := dockerClient.PullImage(test3Image3Name, nil)
+		assert.NoError(t, metadata.Error, "Failed to pull image %s", test3Image3Name)
 	}
 
 	// The same image name used by all tasks in this test
@@ -323,7 +92,7 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	task2.Containers[0].Image = identicalImageName
 	task3.Containers[0].Image = identicalImageName
 
-	err = renameImage(testImage1Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
+	err = renameImage(test3Image1Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
 	assert.NoError(t, err, "Renaming the image failed")
 
 	// start and wait for task1 to be running
@@ -338,7 +107,7 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	imageID1 := imageState1.Image.ImageID
 
 	// Using another image but rename to the same name as task1 for task2
-	err = renameImage(testImage2Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
+	err = renameImage(test3Image2Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
 	require.NoError(t, err, "Renaming the image failed")
 
 	// Start and wait for task2 to be running
@@ -354,7 +123,7 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	require.NotEqual(t, imageID2, imageID1, "The image id in task 2 should be different from image in task 1")
 
 	// Using a different image for task3 and rename it to the same name as task1 and task2
-	err = renameImage(testImage3Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
+	err = renameImage(test3Image3Name, "testimagewithsamenameanddifferentid", "latest", goDockerClient)
 	require.NoError(t, err, "Renaming the image failed")
 
 	// Start and wiat for task3 to be running
@@ -440,13 +209,13 @@ func TestImageWithSameIDAndDifferentNames(t *testing.T) {
 	task3.Containers[0].Image = "testimagewithsameidanddifferentnames-3:latest"
 
 	// Pull the images needed for the test
-	if _, err = dockerClient.InspectImage(testImage1Name); err == docker.ErrNoSuchImage {
-		metadata := dockerClient.PullImage(testImage1Name, nil)
-		assert.NoError(t, metadata.Error, "Failed to pull image %s", testImage1Name)
+	if _, err = dockerClient.InspectImage(test4Image1Name); err == docker.ErrNoSuchImage {
+		metadata := dockerClient.PullImage(test4Image1Name, nil)
+		assert.NoError(t, metadata.Error, "Failed to pull image %s", test4Image1Name)
 	}
 
 	// Using testImage1Name for all the tasks but with different name
-	err = renameImage(testImage1Name, "testimagewithsameidanddifferentnames-1", "latest", goDockerClient)
+	err = renameImage(test4Image1Name, "testimagewithsameidanddifferentnames-1", "latest", goDockerClient)
 	require.NoError(t, err, "Renaming image failed")
 
 	// Start and wait for task1 to be running
@@ -544,55 +313,4 @@ func renameImage(original, repo, tag string, client *docker.Client) error {
 	}
 
 	return nil
-}
-
-func verifyTaskIsCleanedUp(taskName string, taskEngine TaskEngine) error {
-	for i := 0; i < taskCleanupTimeoutSeconds; i++ {
-		_, ok := taskEngine.(*DockerTaskEngine).State().TaskByArn(taskName)
-		if !ok {
-			break
-		}
-		time.Sleep(1 * time.Second)
-		if i == (taskCleanupTimeoutSeconds - 1) {
-			return errors.New("Expected Task to have been swept but was not")
-		}
-	}
-	return nil
-}
-
-func verifyImagesAreRemoved(imageManager *dockerImageManager, imageIDs ...string) error {
-	imagesNotRemovedList := list.New()
-	for _, imageID := range imageIDs {
-		_, ok := imageManager.getImageState(imageID)
-		if ok {
-			imagesNotRemovedList.PushFront(imageID)
-		}
-	}
-	if imagesNotRemovedList.Len() > 0 {
-		return fmt.Errorf("Image states still exist for: %v", imagesNotRemovedList)
-	}
-	return nil
-}
-
-func verifyImagesAreNotRemoved(imageManager *dockerImageManager, imageIDs ...string) error {
-	imagesRemovedList := list.New()
-	for _, imageID := range imageIDs {
-		_, ok := imageManager.getImageState(imageID)
-		if !ok {
-			imagesRemovedList.PushFront(imageID)
-		}
-	}
-	if imagesRemovedList.Len() > 0 {
-		return fmt.Errorf("Could not find images: %v in ImageManager", imagesRemovedList)
-	}
-	return nil
-}
-
-func cleanupImages(imageManager *dockerImageManager) {
-	imageManager.client.RemoveContainer("test1", removeContainerTimeout)
-	imageManager.client.RemoveContainer("test2", removeContainerTimeout)
-	imageManager.client.RemoveContainer("test3", removeContainerTimeout)
-	imageManager.client.RemoveImage(testImage1Name, imageRemovalTimeout)
-	imageManager.client.RemoveImage(testImage2Name, imageRemovalTimeout)
-	imageManager.client.RemoveImage(testImage3Name, imageRemovalTimeout)
 }
