+++ misc/image-cleanup-test-images/linux0.dockerfile
