+++ misc/image-cleanup-test-images/linux1.dockerfile
