+++ misc/image-cleanup-test-images/linux2.dockerfile
